refactor: 파일 구조 변경
- AddCard로 카드 추가 로직 이동 후 라우터 적용
