+++ src/components/AddCard/CardNumberInput.jsx
@@ -1,9 +1,9 @@
 import { Fragment } from "react";
 
-import Input from "./UIComponents/Input/Input";
-import InputField from "./UIComponents/InputField/InputField";
+import Input from "../UIComponents/Input/Input";
+import InputField from "../UIComponents/InputField/InputField";
 
-import { CARD_INFO_RULES } from "../constants/constants";
+import { CARD_INFO_RULES } from "../../constants/constants";
 
 export default function CardNumberInput(props) {
   const { cardNumber, onChange } = props;
