+++ src/components/AddCard/CardPasswordInput.jsx
@@ -1,7 +1,7 @@
-import Input from "./UIComponents/Input/Input";
-import InputField from "./UIComponents/InputField/InputField";
+import Input from "../UIComponents/Input/Input";
+import InputField from "../UIComponents/InputField/InputField";
 
-import { CARD_INFO_RULES } from "../constants/constants";
+import { CARD_INFO_RULES } from "../../constants/constants";
 
 export default function CardPasswordInput(props) {
   const { password, onChange } = props;
