+++ src/components/AddCard/index.jsx
@@ -3,3 +3,4 @@ export { default as CardExpireDateInput } from "./CardExpireDateInput";
 export { default as CardHolderNameInput } from "./CardHolderNameInput";
 export { default as CardSecurityCodeInput } from "./CardSecurityCodeInput";
 export { default as CardPasswordInput } from "./CardPasswordInput";
+export { default as CardInfoForm } from "./CardInfoForm";
